Add scriptable components

+++ engine/script-core/src/Fusion/InternalCalls.cs
@@ -7,18 +7,18 @@ namespace Fusion
 	{
 		#region Entity
 		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal extern static bool Entity_HasComponent(ulong entityID, Type componentType);
+		internal extern static bool Entity_HasComponent(uint entityID, Type componentType);
 		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal extern static ulong Entity_FindEntityByName(string name);
+		internal extern static uint Entity_FindEntityByName(string name);
 		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal extern static object GetScriptInstance(ulong entityID);
+		internal extern static object GetScriptInstance(uint entityID);
 		#endregion
 
 		#region TransformComponent
 		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal extern static void TransformComponent_GetTranslation(ulong entityID, out Vector3 translation);
+		internal extern static void TransformComponent_GetTranslation(uint entityID, out Vector3 translation);
 		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal extern static void TransformComponent_SetTranslation(ulong entityID, ref Vector3 translation);
+		internal extern static void TransformComponent_SetTranslation(uint entityID, ref Vector3 translation);
 		#endregion
 
 		#region Input
